requirements.txt
# === DEPENDENCIAS CORREGIDAS - Red Soluciones ISP ===

# === CORE FRAMEWORK ===
fastapi==0.104.1
uvicorn[standard]==0.24.0

# === MODELS & DATA ===
pydantic==2.11.7
python-multipart==0.0.6

# === HTTP & REQUESTS ===
requests==2.32.4
httpx==0.27.0

# === GOOGLE SERVICES - VERSIONES COMPATIBLES ===
google-generativeai==0.8.5
google-auth==2.40.3
google-api-python-client==2.176.0
gspread==6.0.2
google-auth-oauthlib==1.2.2

# === UTILITIES ===
python-dotenv==1.0.0
tenacity==9.1.2

# === SSL FIX ===
urllib3<2.0.0,>=1.26.0
certifi>=2023.7.22
